satellite_qa_agent_app/requirements.txt
Flask==3.0.3
google-cloud-storage==2.16.0
requests==2.32.3
gunicorn==22.0.0